+++ sslyze/plugins/utils/certificate_utils.py
